requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
